async replies

+++ fictionsuit/config.py
@@ -11,7 +11,7 @@ SERVER = os.getenv("SERVER")
 OAI_MODEL = os.getenv("OAI_MODEL", "gpt-3.5-turbo-0301")
 openai.api_key = os.getenv("OPENAI_API_KEY")
 TEMPERATURE = float(os.getenv("TEMPERATURE", 1))
-MAX_TOKENS = float(os.getenv("MAX_TOKENS", 500))
+MAX_TOKENS = int(os.getenv("MAX_TOKENS", 500))
 
 # Bot specific
 COMMAND_PREFIX = os.getenv("COMMAND_PREFIX", "chat ")
